Fixes #1: Enable custom width cells
Now you can resize cells with TextGrid.ResizeCells(units), this
adds horizontal padding to the cell's content.

Since we need to always have a odd number for the size of the cell,
we implemented a unit system so we don't have to throw any errors
if the user wants to have an even-sized cell since that would not
be symmetric once you add the content.
You will have to redraw the grid each time you resize its cells.